Update node option handling
