+++ src/stores/edit_node.ts
@@ -33,7 +33,7 @@ import { defineStore } from 'pinia'
 import { computed, ref } from 'vue'
 import { useEditorStore } from './editor'
 import { useNodesStore } from './nodes'
-import { getDefaultValue, prettyprint_type, python_builtin_types } from '@/utils'
+import { getDefaultValue, prettyprint_type, python_builtin_types, serializeNodeOptions } from '@/utils'
 
 export enum EditorSelectionSource {
   NONE = 'none',
@@ -312,38 +312,25 @@ export const useEditNodeStore = defineStore('edit_node', () => {
         return
       }
 
-      const option_ref_keys = reference_node.value.options
-        .filter((x) => prettyprint_type(x.serialized_value).startsWith('OptionRef('))
-        .map((x): [string, string] => [
-          x.key,
-          prettyprint_type(x.serialized_value).substring(
-            'OptionRef('.length,
-            prettyprint_type(x.serialized_value).length - 1
+      function findOptionRefs(ref_list: NodeData[]): [string, string][] {
+        return ref_list.filter((x) => 
+            prettyprint_type(x.serialized_value).startsWith('OptionRef(')
           )
-        ])
-        .filter((x) => x[1] === key)
+          .map((x): [string, string] => [
+            x.key,
+            prettyprint_type(x.serialized_value).substring(
+              'OptionRef('.length,
+              prettyprint_type(x.serialized_value).length - 1
+            )
+          ])
+          .filter((x) => x[1] === key)
+      }
 
-      const input_ref_keys = reference_node.value.inputs
-        .filter((x) => prettyprint_type(x.serialized_value).startsWith('OptionRef('))
-        .map((x): [string, string] => [
-          x.key,
-          prettyprint_type(x.serialized_value).substring(
-            'OptionRef('.length,
-            prettyprint_type(x.serialized_value).length - 1
-          )
-        ])
-        .filter((x) => x[1] === key)
+      const option_ref_keys = findOptionRefs(reference_node.value.options)
 
-      const output_ref_keys = reference_node.value.outputs
-        .filter((x) => prettyprint_type(x.serialized_value).startsWith('OptionRef('))
-        .map((x): [string, string] => [
-          x.key,
-          prettyprint_type(x.serialized_value).substring(
-            'OptionRef('.length,
-            prettyprint_type(x.serialized_value).length - 1
-          )
-        ])
-        .filter((x) => x[1] === key)
+      const input_ref_keys = findOptionRefs(reference_node.value.inputs)
+
+      const output_ref_keys = findOptionRefs(reference_node.value.outputs)
 
       new_node_options.value = new_node_options.value.map(map_fun)
       function resolve_refs(refs: [string, string][], current_item: ParamData): ParamData {
@@ -385,32 +372,6 @@ export const useEditNodeStore = defineStore('edit_node', () => {
     }
   }
 
-  function serializeNodeOptions(): NodeData[] {
-    return new_node_options.value.map((x) => {
-      const option: NodeData = {
-        key: x.key,
-        serialized_value: '',
-        serialized_type: '' // This is left blank intentionally
-      }
-      if (x.value.type === 'type') {
-        if (python_builtin_types.indexOf(x.value.value as string) >= 0) {
-          x.value.value = 'builtins.' + x.value.value;
-        }
-        option.serialized_value = JSON.stringify({
-          'py/type': x.value.value
-        })
-      } else if (x.value.type.startsWith('__')) {
-        //TODO This should be changed to not generate "bad" defaults
-        const val = x.value.value as PyObject
-        val['py/object'] = x.value.type.substring('__'.length)
-        option.serialized_value = JSON.stringify(x.value.value)
-      } else {
-        option.serialized_value = JSON.stringify(x.value.value)
-      }
-      return option
-    })
-  }
-
   function buildNodeMsg(): NodeMsg {
     return {
       module: new_node_module.value,
@@ -418,7 +379,7 @@ export const useEditNodeStore = defineStore('edit_node', () => {
       name: new_node_name.value,
       max_children: 0,
       child_names: [],
-      options: serializeNodeOptions(),
+      options: serializeNodeOptions(new_node_options.value),
       inputs: [],
       outputs: [],
       version: '',
@@ -454,7 +415,6 @@ export const useEditNodeStore = defineStore('edit_node', () => {
     changeNodeName,
     changeNodeClass,
     updateParamValue,
-    serializeNodeOptions,
     buildNodeMsg,
   }
 })
